# config/seng365-2018_auction_0.0.7_swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "info": {
    "description": "",
    "version": "0.0.6",
    "title": "Auction API for SENG365 2018 Assignments"
  },
...
  "paths": {
...
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Log in user by username or email",
        "description": "One of username and email is required",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "The username for login",
            "schema": {
              "type" : "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "description": "The users email for login",
            "schema": {
              "$ref" : "#/components/schemas/email"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "The password for login in clear text",
            "required": true,
            "schema": {
              "$ref" : "#/components/schemas/password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/logInResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username/email/password supplied"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "logInResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "the id of the logged in user"
          },
          "token": {
            "type": "string",
            "description": "a token to be used for future calls"
          }
        }
      },
...
      "email": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "password": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
...
    },
    "securitySchemes": {
      "User": {
        "type": "apiKey",
        "name": "X-Authorization",
        "in": "header"
      }
    }
  }
}